fetch/metadata/generated/element-video-poster.https.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-video-poster.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "video" element "poster"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  const params = {
    body: `
      <svg xmlns="http://www.w3.org/2000/svg" width="123" height="123">
        <rect fill="lime" width="123" height="123"/>
      </svg>
    `,
    mime: 'image/svg+xml'
  };

  function induceRequest(t, url) {
    var video = document.createElement('video');
    video.setAttribute('poster', url);
    document.body.appendChild(video);

    const poll = () => {
      if (video.clientWidth === 123) {
        return;
      }

      return new Promise((resolve) => t.step_timeout(resolve, 0))
        .then(poll);
    };
    t.add_cleanup(() => video.remove());

    return poll();
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsOrigin'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-origin']);
      });
  }, 'sec-fetch-frame-ancestors - Same origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsCrossSite'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['cross-site']);
      });
  }, 'sec-fetch-frame-ancestors - Cross-site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsSameSite'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-site']);
      });
  }, 'sec-fetch-frame-ancestors - Same site');
  </script>
  </body>
</html>
